Fixed cool/interesting 32-bit bugs

+++ src/ArrayStringLong.cpp
@@ -96,7 +96,7 @@ void ArrayStringLong::Delete(size_t ndx) {
 
 	m_blob.Delete(start, end);
 	m_offsets.Delete(ndx);
-	m_offsets.Adjust(ndx, start - end);
+	m_offsets.Adjust(ndx, (int64_t)start - end);
 }
 
 void ArrayStringLong::Clear() {
